Reworked lazy values completely
Previously the element type of the lazyly loaded array was discovered
at runtime using instanceof checks when the value was loaded. Now the
LazyArrayValue is split into one concrete class per supported primitive
instead, similarly to the direct value classes. This allows us to
specialize and to less casting, and have more compile time type safety.

Also removed defensive copying of arrays! This is not needed as we are
no longer exposing the inner array objects to the outside.