Scanning and Jumping PageCacheBackedCursor modes
The PageCacheBackedCursor has two modes of operation: Scanning and Jumping.

The cursor starts in the Unbound state. From there it can be initialized as
either a scanning cursor or a jumping one. Once in a certain mode, only the
traversal operations particular to that mode can be called. It is the
responsibility of extending classed to ensure this. On `close()` the cursor
will revert to the unbound state, and can be reused in either of the
two modes.